# Redesign import filter system for acceptable performance.