--- a/library/libobject.ml
+++ b/library/libobject.ml
@@ -18,30 +18,22 @@ type 'a substitutivity =
 
 type object_name = Libnames.full_path * Names.KerName.t
 
-module NSet = Globnames.ExtRefSet
-
 type open_filter =
   | Unfiltered
-  | Names of NSet.t
+
 
 let simple_open f filter i o = match filter with
   | Unfiltered -> f i o
-  | Names _ -> ()
+
 
 let filter_and f1 f2 = match f1, f2 with
-  | Unfiltered, f | f, Unfiltered -> Some f
-  | Names n1, Names n2 ->
-    let n = NSet.inter n1 n2 in
-    if NSet.is_empty n then None
-    else Some (Names n)
+  | Unfiltered, f -> Some f
 
 let filter_or f1 f2 = match f1, f2 with
-  | Unfiltered, f | f, Unfiltered -> Unfiltered
-  | Names n1, Names n2 -> Names (NSet.union n1 n2)
+  | Unfiltered, f -> Unfiltered
 
 let in_filter_ref gr = function
   | Unfiltered -> true
-  | Names ns -> NSet.mem (Globnames.TrueGlobal gr) ns
 
 type 'a object_declaration = {
   object_name : string;