Fix #14570

Instead of running through all objects and testing that they match the
filter, we lookup the specific things we want to import.

Tested with a file all_stdlib.v containing
~~~coq
Require Export Reals.ROrderedType
Reals.Alembert
...

Definition toimport := Prop.
~~~
(similar to the one used in test-suite/misc/universes but with Export
+ a dummy definition)

then
~~~coq
Require all_stdlib.

Time Import all_stdlib(toimport).
~~~
Unfiltered import before and after the patch takes between 0.2s and 0.3s.
Filtered import:
- before the patch: 9s
- after the patch: 0s
- with another patch which tried to be smarter in
  Declaremods.open_module: 0.9s (this was not satisfying so I made the
  present redesign)

Dummy filters are still present for code compatibility, they may be
removed in a future commit, or reused for non-names types of
filters (eg filtering by kind of object (notations,coercions,etc)
should be faster as there is no need to compare many strings).